requirements.txt
streamlit
pytesseract
Pillow
pillow-heif
